ParkApp v0.0.1 Parkapp.Auth.Devices

The Devices module contains the Device/DB communication related methods.

Functions

create_device(attrs)

Creates a device.

Examples

iex> create_device(%{field: value})
{:ok, %Device{}}

iex> create_device(%{field: bad_value})
{:error, %Ecto.Changeset{}}

get_device!(device_id)

Gets a single device.

Raises Ecto.NoResultsError if the device does not exist

Examples

iex> get_device(123)
%Device{}

iex> get_device(456)
Ecto.NoResultsError

get_device(device_id)

Gets a single device.

Returns nil if there is no device with the given id

Examples

iex> get_device(123)
%Device{}

iex> get_device(456)
nil

get_device_field(device_id, field)

Gets a field from a single device.

Returns nil if there is no device with the given id

Examples

iex> get_device_field(123, :field)
value

iex> get_device_field(456, :field)
nil

update_device_base(device, attrs)

Updates a device base.

Examples

iex> update_device_base(device, %{field: new_value})
{:ok, %Device{}}

iex> update_device_base(device, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_device_secret(device, attrs)

Updates a device secret.

Examples

iex> update_device_secret(device, %{field: new_value})
{:ok, %Device{}}

iex> update_device_secret(device, %{field: bad_value})
{:error, %Ecto.Changeset{}}