ParkApp v0.0.1 Parkapp.ReservationsContext

The Reservations context.

Functions

change_reservation(reservation)

Returns an %Ecto.Changeset{} for tracking reservation changes.

Examples

iex> change_reservation(reservation)
%Ecto.Changeset{source: %Reservation{}}

change_reservation_status_history(reservation_status_history)

Returns an %Ecto.Changeset{} for tracking reservation_status_history changes.

Examples

iex> change_reservation_status_history(reservation_status_history)
%Ecto.Changeset{source: %ReservationStatusHistory{}}

create_reservation(attrs \\ %{})

Creates a reservation.

Examples

iex> create_reservation(%{field: value})
{:ok, %Reservation{}}

iex> create_reservation(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_reservation_initial_state(attrs)

create_reservation_initial_state(Map) ::
  {:ok, Parkapp.Reservations.Reservation} | {:error, Changeset} | nil

Creates a reservation with the given attributes in the open state. Creates the first entry in the status history.

create_reservation_status_history(attrs \\ %{})

Creates a reservation_status_history.

Examples

iex> create_reservation_status_history(%{field: value})
{:ok, %ReservationStatusHistory{}}

iex> create_reservation_status_history(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_reservation_status_history(reservation_id, previous_status, next_status)

Creates a reservation_status_history.

deactivate_reservation_status_history(reservation_id)

deactivate_reservation_status_history(Integer) ::
  {:ok, Parkapp.Reservations.ReservationStatusHistory}
  | {:error, Changeset}
  | nil

Gets the active history for the given reservation and deactivates it.

delete_reservation(reservation)

Deletes a Reservation.

Examples

iex> delete_reservation(reservation)
{:ok, %Reservation{}}

iex> delete_reservation(reservation)
{:error, %Ecto.Changeset{}}

delete_reservation_status_history(reservation_status_history)

Deletes a ReservationStatusHistory.

Examples

iex> delete_reservation_status_history(reservation_status_history)
{:ok, %ReservationStatusHistory{}}

iex> delete_reservation_status_history(reservation_status_history)
{:error, %Ecto.Changeset{}}

get_active_reservation_history(reservation_id)

get_active_reservation_history(Integer) ::
  Parkapp.Reservations.ReservationStatusHistory | nil

Retrieves from the database the active history for the given reservation

get_current_reservation(device_id)

get_current_reservation(Integer) :: Parkapp.Reservations.Reservation | nil

Retrives the current reservation for the given device. It makes sure the status history is correct and ignores the closed reservations

get_reservation!(id)

Gets a single reservation.

Raises Ecto.NoResultsError if the Reservation does not exist.

Examples

iex> get_reservation!(123)
%Reservation{}

iex> get_reservation!(456)
** (Ecto.NoResultsError)

get_reservation_status!(id)

Gets a single reservation_status.

Raises Ecto.NoResultsError if the Reservation status does not exist.

Examples

iex> get_reservation_status!(123)
%ReservationStatus{}

iex> get_reservation_status!(456)
** (Ecto.NoResultsError)

get_reservation_status_history!(id)

Gets a single reservation_status_history.

Raises Ecto.NoResultsError if the Reservation status history does not exist.

Examples

iex> get_reservation_status_history!(123)
%ReservationStatusHistory{}

iex> get_reservation_status_history!(456)
** (Ecto.NoResultsError)

get_reservation_type!(id)

Gets a single reservation_type.

Raises Ecto.NoResultsError if the Reservation type does not exist.

Examples

iex> get_reservation_type!(123)
%ReservationType{}

iex> get_reservation_type!(456)
** (Ecto.NoResultsError)

list_reservation_status()

Returns the list of reservation_status.

Examples

iex> list_reservation_status()
[%ReservationStatus{}, ...]

list_reservation_types()

Returns the list of reservation_types.

Examples

iex> list_reservation_types()
[%ReservationType{}, ...]

list_reservations()

Returns the list of reservations.

Examples

iex> list_reservations()
[%Reservation{}, ...]

list_revervation_status_history()

Returns the list of revervation_status_history.

Examples

iex> list_revervation_status_history()
[%ReservationStatusHistory{}, ...]

move_to_closed_state(reservation)

move_to_closed_state(Parkapp.Reservations.Reservation) ::
  {:ok, Parkapp.Reservations.Reservation} | {:error, Changeset} | nil

Updates the reservation to the closed state. Creates the sixth and final entry in the status history. Deactivates the previous one.

move_to_external_payment_state(reservation)

move_to_external_payment_state(Parkapp.Reservations.Reservation) ::
  {:ok, Parkapp.Reservations.Reservation} | {:error, Changeset} | nil

Updates the reservation to the external payment state. Creates the forth entry in the status history. Deactivates the previous one.

move_to_in_park_state(reservation)

move_to_in_park_state(Parkapp.Reservations.Reservation) ::
  {:ok, Parkapp.Reservations.Reservation} | {:error, Changeset} | nil

Updates the reservation to the in park state. Creates the second entry in the status history. Deactivates the previous one.

move_to_payment1_state(reservation, context_token, amount)

move_to_payment1_state(Parkapp.Reservations.Reservation, String, String) ::
  {:ok, Parkapp.Reservations.Reservation} | {:error, Changeset} | nil

Updates the reservation to the payment1 state. Sets the context_token and amount. Creates the third entry in the status history. Deactivates the previous one.

move_to_payment2_state(reservation)

move_to_payment2_state(Parkapp.Reservations.Reservation) ::
  {:ok, Parkapp.Reservations.Reservation} | {:error, Changeset} | nil

Updates the reservation to the payment2 state. Creates the fifth entry in the status history. Deactivates the previous one.

update_reservation(reservation, attrs)

Updates a reservation.

Examples

iex> update_reservation(reservation, %{field: new_value})
{:ok, %Reservation{}}

iex> update_reservation(reservation, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_reservation_after_payment(reservation, attrs)

Updates a reservation using the payment changeset.

Examples

iex> update_reservation(reservation, %{field: new_value})
{:ok, %Reservation{}}

iex> update_reservation(reservation, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_reservation_status_history(reservation_status_history, attrs)

Updates a reservation_status_history.

Examples

iex> update_reservation_status_history(reservation_status_history, %{field: new_value})
{:ok, %ReservationStatusHistory{}}

iex> update_reservation_status_history(reservation_status_history, %{field: bad_value})
{:error, %Ecto.Changeset{}}