ParkApp v0.0.1 ParkappWeb.AuthenticationController

The AuthenticationController module takes care of the handshake process between the server and the mobile-client.

This process is done in 2 consecutive steps: registration and authentication.

The registration process envolves the mobile-client sending his device id and his public key to the server, which will be stored in the database.

The authentication process is a challenge-response process, and has 2 phases. In the first phase, the mobile-client tells the server he wants to authenticate, and the server answers with a secret (a random string) encrypted with the client’s public key. The client decrypts the secret. In the second phase, the client encrypts the secret again, but with the server public key, and sends it to the server. The server then decrypts this secret, and if it matches the secret generated in the first phase, the server sends a token to the client, which he will use in all other requests to prove that he can access the server’s endpoints.

Functions

authenticate_phase1(conn, arg2)

First authentication phase. Associates an encrypted secret with the given device

authenticate_phase2(conn, arg2)

Second authentication phase. Generates a JWT token for the given device.

logout(conn, _)

register(conn, arg2)

Registers a new device

If the device already exists just returns :ok

verify_token(conn, arg2)

Checks if the connection has a legit token associated with it.