ParkApp v0.0.1 ParkappWeb.Utils

The Utils module contains generic methods to be used from the ParkappWeb

Functions

format_float_decimal_places(float, num_decimal_places)

format_float_decimal_places(Float, Integer) :: String

Formats a flot with the given number of decimal places. If the value is not a flot, it is returned as a string.

parse_string_to_float(value)

parse_string_to_float(String) :: Float | nil

Parses the given string to float. It is not a string, nil is returned

random_string(length)

random_string(Integer) :: String

Generates a random string of the given length